requirements.txt
transformers==4.50.0
torch==2.6.0
pillow==10.3.0
scikit-learn==1.6.1
matplotlib==3.10.0
nltk==3.9.1
openai==1.65.1
python-dotenv==1.0.1
tqdm==4.67.1
accelerate==1.4.0
pandas==2.2.3
umap-learn==0.5.7
tabulate==0.9.0
hdbscan==0.8.40
vertexai==1.71.1